Fix Permissions for admin/user controllers

+++ app/Http/Controllers/Admin/User/NotesController.php
@@ -8,11 +8,14 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Contracts\View\View;
+use Illuminate\Support\Facades\Gate;
 
 class NotesController extends Controller
 {
     public function store(Request $request, User $user): RedirectResponse
     {
+
+        abort_if(Gate::denies('user_manage_notes'), 403);
 
         $input = $request->validate([
             'note' => 'required',
@@ -27,6 +30,8 @@ class NotesController extends Controller
 
     public function destroy(User $user, UserNotes $userNotes): RedirectResponse
     {
+        abort_if(Gate::denies('user_manage_notes'), 403);
+
         $userNotes->delete();
         return redirect()->route('admin.users.show', $user->id)->with('alerts', [['message' => 'Note deleted.', 'level' => 'success']]);;
     }
